Follow best practices for prod deployment of React, compiling out all calls to process.env.NODE_ENV
