+++ package.json
@@ -4,11 +4,11 @@
   "description": "rax-ssr",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon app.js",
+    "start": "NODE_ENV=production nodemon app.js",
     "webpack:client": "webpack --config webpack.client.config.js",
     "webpack:server": "webpack --config webpack.server.config.js",
     "webpack": "npm run webpack:client && npm run webpack:server",
-    "benchmark": "bash ./benchmarks/index.sh"
+    "benchmark": "NODE_ENV=production bash ./benchmarks/index.sh"
   },
   "keywords": [
     "ssr",
@@ -40,6 +40,7 @@
   "devDependencies": {
     "babel-cli": "^6.18.0",
     "babel-loader": "^6.2.8",
+    "babel-plugin-transform-runtime": "^6.15.0",
     "babel-plugin-transform-vue-jsx": "^3.2.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
